+++ cmd/cli.go
@@ -1,0 +1,7 @@
+package main
+
+import "nuvlaedge-go/cli/cmd"
+
+func main() {
+	cmd.Execute()
+}